Add Transaction.result_dict to access result data
The transaction result was stored as a Python style pretty-printed
string in a text field. The data could be read by humans in the admin,
but could not be accessed in code.

Instead, the result is now stored as pretty-printed JSON. It is still
human readable, but now also accessible from code. A
`Transaction.result_dict` property has been added that allows
getting/setting the `result` attribute as a dict.