Use `QuerySet.as_manager` instead of custom manager
The custom manager used the deprecated and removed
`Manager.get_query_set` method, instead of the renamed
`Manager.get_queryset` method. Instead of fixing this, I replaced it
with a `QuerySet` which can call `.filter` directly, and used the newish
`QuerySet.as_manager` classmethod to make the manager.

This requires dropping support for Django 1.6.